authentication/requirements.txt
flask ~= 2.2.3
grpcio-tools ~= 1.30
Jinja2 ~= 3.1.2
Authlib
requests
python-dotenv>=0.19.2
prometheus_client
psutil